enhance: export custom nextjs/socket.io server
Package and export a function for creating the custom combination NextJS/socket.io
server so the UI can be more easily imported by other projects.
